backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
